FIX: Windows compilation
Windows compilation broke with the more restrictive compiler flags
and folder restructuring. Fix the errors so that it compiles again.

Make stepTime float instead of double.

Update the README with feature images and step-by-step for Windows.

+++ include/app/objects/actuators/WheelMotor.h
@@ -51,7 +51,7 @@ public:
 
     void setVoltageIn(float voltage);
     void setDutyCycle(float dutyCycle);
-    void onFixedUpdate(double stepTime) override;
+    void onFixedUpdate(float stepTime) override;
     float *getVoltageLine();
     const Body2D *getBody() const { return m_body2D; }
