+++ include/engine/renderer/VertexBuffer.h
@@ -1,6 +1,7 @@
 #ifndef VERTEX_BUFFER_H_
 #define VERTEX_BUFFER_H_
 
+#include <cstddef>
 /**
  * Wrapper around OpenGL vertex buffer object (VBO). A VBO holds vertex data such as
  * position and texture coordinates.
@@ -9,9 +10,9 @@ class VertexBuffer
 {
 public:
     enum class DrawType { Static, Dynamic };
-    VertexBuffer(const void* data, unsigned int size, DrawType drawType);
+    VertexBuffer(const void* data, size_t size, DrawType drawType);
     ~VertexBuffer();
-    void updateData(const void *data, unsigned int size);
+    void updateData(const void *data, size_t size);
 
     void bind() const;
     void unbind() const;
@@ -20,7 +21,7 @@ private:
     int glDrawType(DrawType drawType);
 
     unsigned int m_id;
-    const unsigned int m_size;
+    const size_t m_size;
     const int m_glDrawType;
 };
