+++ testapp/scenes/WheelMotorTestScene.cpp
@@ -29,7 +29,7 @@ namespace {
                 }
             }
         }
-        void onFixedUpdate(double stepTime) override
+        void onFixedUpdate(float stepTime) override
         {
             (void)stepTime;
         }
